Add pre-push driver hook
This is a "driver" hook because it sources other scripts (which I'm
going to call "auxiliary hooks") to do all the real work. It should be
copied into a repository's .git/hooks directory and edited to source the
auxiliary hooks that are appropriate for that repository.

I haven't implemented any auxiliary hooks yet, but they will be located
in the githooks working copy and, unlike the drivers, will not have to
be copied to repositories that use them.