Redesign and rewrite entire system
(Not that there was much to rewrite.)

Previously, `pre-push` (the only driver hook so far -- the holotype, if
you will) was designed to be copied into a repository's hooks directory
and edited to activate auxiliary hooks. It was written in POSIX shell
and sourced the auxiliary hooks, so those had to be written in POSIX
shell too, and they had to tread carefully because they shared internal
state with other auxiliary hooks. Future driver hooks would have been
conceptually similar to `pre-push` but implemented separately because
they would have had their own lists of auxiliary hooks.

Now, there is a single driver hook called `omnihook` which should be
placed into a repository's hooks directory and given the name of a Git
hook 'FOO'. All executables in the hooks directory whose names begin
with 'FOO-' are considered auxiliary hooks and are automatically run by
the driver hook, which no longer needs to be modified and can be linked
instead of copied. (This obviates the purpose of using '$Id$'.) The
driver hook can be used as *any* Git hook, so future non-`pre-push`
auxiliary hooks will not need separate driver hooks.

Because auxiliary hooks are no longer sourced, they can be any sort of
executable -- indeed, the driver hook expects them to be usable as
standalone hooks. The pre-push-no-WIP auxiliary hook is now a proper
POSIX shell script, and pre-push-require-sigs has been reimplemented as
an awk script.

Changes that aren't really related to the design of the system:
    - pre-push-no-WIP no longer recognizes bare keywords if they are
      followed by a letter. Overly enthusiastic commit messages
      beginning with "REWORDED" and "SQUASHED" are okay now!
    - pre-push-no-WIP now recognizes "NOCOMMIT" as a WIP keyword.
    - pre-push-no-WIP now sources Git's git-sh-setup library, which has
      some useful stuff. (It doesn't USE any of that stuff, but still.)
    - pre-push-require-sigs now accepts all "good" signatures, even
      those that have unknown validity, are expired, or were made by
      expired or revoked certificates.
    - Both hooks now list the commits that prompted push cancellation.