require-sigs: Rewrite in POSIX shell
I'd hoped that the awk rewrite [*] would yield a significant improvement
over shell. Unfortunately, due to the natures of the input and task, the
awk implementation was just as complex as the shell implementation with
which I am replacing it now, and it had its own peculiarities that had
to be worked around and explained. I didn't even benefit from awk's
speed because I rarely push more than a handful of commits at a time.

The shell replacement is just as portable, feels more natural in some
ways, and is more amenable to additional `git` calls, which might be
necessary for future improvements.

(Incidentally, fix a bug wherein the "blocked push to ..." error message
would be printed before the first range of commits only, instead of
before each range.)

    [*]: f9c0234 (Redesign and rewrite entire system, 2020-03-16)