no-WIP: Clarify logic for changing the exit status
I was never satisfied with the previous logic. If there were no WIP
commits in a given range, then the `if read -r` command -- and therefore
the whole pipeline -- would exit with a zero status. This wasn't
a problem per se, but it meant that when there _were_ WIP commits, the
pipeline had to exit with a _nonzero_ status. Ending the `if read -r`
command list with `false` worked fine but looked weird, and it always
took me a few moments to relearn how the control flow worked.

Running the second pipeline stage in an explicit subshell strikes me as
the cleanest solution. It enables (portably) exiting the pipeline early,
with a nonzero status, if there are no WIP commits in the current range.
(I'm a big fan of early exits.) Otherwise, the subshell can carry on and
ideally finish with a zero status, without requiring a `true` or `:` to
force it. This exit-status switcheroo is easily accommodated by changing
`||` to `&&` after the pipeline. I find the result much more natural.

Tweak the comments while I'm at it.

(Background: It would not be reliable to assign to `status` from within
a pipeline stage because they often run in implicit subshells [*]. Thus,
`status` can only be changed from outside the pipeline, and the pipeline
has to influence the assignment through external means such as exit
status.)

    [*]: https://mywiki.wooledge.org/BashFAQ/024