omnihook: Invoke each hook once, passing all stdin
Previously, we invoked each hook once per line of standard input and
drip-fed it that line. This was straightforward and avoided the need to
pool the input, but it altered the effective behavior of hooks designed
to process all standard input before exiting (like my own!). Such hooks
essentially sprouted fail-fast behavior spontaneously because they only
saw one line of standard input per invocation.

Now, if expecting standard input, save it to a temporary file, then
invoke each hook only once, with the file on standard input; this avoids
the errant fail-fast behavior. Increased access to persistent storage
might slow things down, but in practice that might not be perceptible.
Likewise for any increase in required storage. (As a bonus, for H hooks
and L lines of standard input, we now spawn H direct child processes,
down from H × L.)

The alternative was to store the standard input in a variable and pass
it to the hooks in a here-document. The code would have been simpler,
but memory usage would have been less predictable. It might have been
faster, but many shells implement here-documents via temporary files
anyway and would have incurred a similar access penalty.