Tell ShellCheck to ignore `git-sh-setup` explicitly
I don't want ShellCheck to try looking at `git-sh-setup` because it
isn't my code, but the "disable" directive in .shellcheckrc is overly
broad.  Also, it can be bypassed with `shellcheck --norc` or by running
ShellCheck from a different directory.

Use `source=/dev/null` directives directly above the `. git-sh-setup`
commands to target the effects and avoid depending on an external file.