Rewrite hook summaries
It is silly to provide a one-sentence summary in the READMEs and direct
the reader to the source code for more detail because the descriptions
in the source code comprise only a few sentences themselves.  Expand the
summaries and incorporate them into the source code for consistency.

Add detail to the description in omnihook.m4 and align it more closely
with the READMEs' "Usage" sections.

Add some unnecessary borders around the source code descriptions because
I kind of like how they look.