no-WIP: Ensure rc gets set to 1 if there are WIPs
Back in 3fdc9d3 (no-WIP: Clarify logic for changing the exit status,
2020-06-15), I hinted at a potential issue but wrongly dismissed it:

    [The] subshell can carry on and ideally finish with a zero status,
    without requiring a `true` or `:` to force it.

If the "if" pipeline fails, then rc is not set to 1.  This should only
happen if "git rev-list" finds no WIPs and outputs nothing, causing
read(1) to fail.  However, it can also happen if "git rev-list" *does*
find WIPs, but the last command in the subshell fails.  This is unlikely
yet unacceptable; the commands in the subshell are purely cosmetic and
must not influence the hook's result.

End the subshell with a ":" command to ensure that it always succeeds
and rc always gets set to 1 if "git rev-list" outputs anything.