no-WIP: Match WIP strings at start of message only
The hook finds WIP commits using "git rev-list --grep", which (as far as
I can tell) is unable to limit matches to the very beginnings of commit
messages.  So it blocks commits with messages containing a WIP string at
the beginning of *any* line, not just the first one.  This was never the
intended behavior, but I'd never gotten around to fixing it.  Until now!

Replace the --grep options with a grep(1) command that does what I want.