omnihook.m4: Rename USE_STDIN_CACHE to READS_STDIN
The name "USE_STDIN_CACHE" is not wrong, but it's not precise either;
one could reasonably assume that the macro just toggles between using
a cache and not using a cache.  It's really meant to distinguish the
hooks that read standard input from the ones that don't; cache use is
downstream of that.

Rename the macro to READS_STDIN to reflect its purpose more precisely.