no-WIP: Match WIP prefixes case-sensitively
Matching ignores case to be forgiving of typos.  However, an upcoming
commit will add support for "git rebase --interactive --autosquash"
prefixes ("amend! ", "fixup! ", and "squash! "), which will be matched
case-sensitively (because Git does so).  I think it would be confusing
if (and annoying to document that) the hook were case-sensitive for the
autosquash prefixes but case-insensitive for the others.

Do all matching case-sensitively to avoid future inconsistency.