pre-push-*: Redirect standard output more locally
There's nothing wrong with redirecting standard output for the whole
script, but I don't love that the redirection command is so distant and
easy to forget about.

Redirect the standard output of the "if" subshell to standard error and
remove the global redirection.  The only commands that actually generate
output are in that subshell, so the fact that they are redirected is
more obvious.  Future commands outside the subshell will have to be
addressed separately, but I don't anticipate any such commands.