no-WIP: Require a space after WIP prefixes
The non-autosquash prefixes match even if they are immediately adjacent
to the rest of the commit title (e.g., "[WIP]fix something").  This
contrasts with the autosquash prefixes, which are only recognized by Git
(and this hook) when followed by spaces (not tabs, interestingly).

For consistency, require a space after WIP prefixes, as with autosquash
prefixes.  As an edge case, continue to accept WIP prefixes that make up
the entire title because I don't know how else to handle them.

In README.rst, stop using ``...`` syntax for the WIP prefixes because
I can't get it to work with the new trailing space.