pre-push-*: Accommodate SHA-256 object hashes
The all-zeros object name is assumed to be a string of 40 zeros -- i.e.,
a SHA-1 digest expressed in hexadecimal.  However, recent versions of
Git can hash objects using SHA-256, for which the all-zeros object name
is a string of 64 zeros.

Test for all-zeros object names without assuming any particular length.
This accommodates SHA-256 and potential future alternatives, as long as
all-zeros object names continue to be used in the same way.

Rename *_sha1 variables to *_obj because they are no longer limited to
SHA-1 digests.